hangar + vehicule v2

+++ app/Domaine/Business/Materiel/EmplacementBusiness.php
@@ -4,6 +4,7 @@ namespace App\Domaine\Business\Materiel;
 
 use App\Domaine\Exceptions\ArrayException;
 use App\Models\Article;
+use App\Models\Hangar;
 use DB;
 use \Illuminate\Database\Eloquent\Collection;
 use App\Models\Emplacement;
@@ -26,7 +27,7 @@ class EmplacementBusiness
    */
   public static function listEmplacements()
   {
-    return Emplacement::all();
+    return Emplacement::with(['article', 'hangar'])->get();
   }
 
   /**
@@ -36,12 +37,14 @@ class EmplacementBusiness
    */
   public static function getEmplacement($id)
   {
-    return Emplacement::find($id);
+    return Emplacement::with(['article', 'hangar'])->find($id);
   }
 
   /**
    * Create a new emplacement
-   * @param array $emplacement #emplacement_new Properties of the new emplacement
+   * @param array $emplacement #emplacement_new Properties of the new emplacement, avec
+   * éventuellement un sous-objet hangar (rue, no_rue, localite_id) si cet emplacement
+   * représente un hangar (bâtiment)
    * @return #idobj ID of the created emplacement
    */
   public static function createEmplacement($emplacement)
@@ -49,7 +52,38 @@ class EmplacementBusiness
     $order = DB::table('emplacements')->max('id');
     $emplacement['remarque'] ??= '';
     $emplacement['tri'] = ($order ?? 0) + 1;
-    return Emplacement::create($emplacement);
+
+    $hangar = $emplacement['hangar'] ?? null;
+    unset($emplacement['hangar']);
+
+    $created = Emplacement::create($emplacement);
+
+    if ($hangar !== null) {
+      Hangar::create(['id' => $created->id, ...$hangar]);
+    }
+
+    return Emplacement::with(['article', 'hangar'])->find($created->id);
+  }
+
+  /**
+   * Create the emplacement representing an article (ex: the inside of a vehicule)
+   * @param Article $article Article represented by the new emplacement
+   * @param array $emplacementData couleur_id, parent_id, est_etiquete, est_compartimentable
+   * @return Emplacement
+   */
+  public static function createEmplacementPourArticle(Article $article, array $emplacementData): Emplacement
+  {
+    $order = DB::table('emplacements')->max('id');
+    return Emplacement::create([
+      'designation' => $article->designation,
+      'remarque' => $article->remarque,
+      'tri' => ($order ?? 0) + 1,
+      'couleur_id' => $emplacementData['couleur_id'],
+      'parent_id' => $emplacementData['parent_id'] ?? null,
+      'est_etiquete' => $emplacementData['est_etiquete'] ?? false,
+      'est_compartimentable' => $emplacementData['est_compartimentable'] ?? false,
+      'article_id' => $article->id,
+    ]);
   }
 
   /**
@@ -59,10 +93,25 @@ class EmplacementBusiness
    */
   public static function editEmplacement($id, $data)
   {
+    if (Emplacement::find($id)->article_id !== null) {
+      throw new ArrayException([], "Cet emplacement est géré depuis la fiche du véhicule lié");
+    }
     $data['remarque'] ??= '';
+
+    // Un sous-objet hangar n'est présent que lorsque l'édition vient de la gestion des
+    // hangars (ModalHangar) : son absence (édition via la gestion générique des
+    // emplacements) ne doit pas effacer un hangar existant.
+    $hangar = $data['hangar'] ?? null;
+    unset($data['hangar']);
+
     Emplacement::whereId($id)
       ->update($data);
-    return Emplacement::find($id);
+
+    if ($hangar !== null) {
+      Hangar::updateOrCreate(['id' => $id], $hangar);
+    }
+
+    return Emplacement::with(['article', 'hangar'])->find($id);
   }
 
   /**
@@ -74,6 +123,12 @@ class EmplacementBusiness
   {
     if (Article::where('emplacement_id', $id)->exists()) {
       throw new ArrayException([], "Veuillez d'abord supprimer les article de cet emplacement");
+    }
+    if (Emplacement::where('parent_id', $id)->exists()) {
+      throw new ArrayException([], "Veuillez d'abord supprimer ou déplacer les sous-emplacements de cet emplacement");
+    }
+    if (Emplacement::find($id)->article_id !== null) {
+      throw new ArrayException([], "Cet emplacement est lié à un véhicule, il ne peut être supprimé qu'en supprimant le véhicule");
     }
     return Emplacement::whereId($id)->delete();
   }
